requirements.txt
Waifulabs
Pillow
psutil